Set the status light's look for each of the dock's states
The dock's states are now Starting, No Wi-Fi, Post failed, Sensor
missing and Well, shown highest first, and the server's dock.led block
gives each a look: off, solid, a pulse or a flash, repeating every 0.25
to 10 s. The defaults keep a slow pulse while the dock works. Updating
keeps its own pulse. The Dock tab sets the looks under Status light,
and shows a look's interval only for a pulse or a flash.

+++ include/net/ClientStatus.h
@@ -32,7 +32,7 @@ struct ClientStatus {
     uint16_t    bsecSampleS;     // seconds between its samples
     const char* settingsVersion; // the board settings it runs, "" before any; null on a
                                  // board without settings, which sends neither block
-    uint8_t     settingsRefused; // the keys of them it refused, as BoardSettings.h's bits
+    uint32_t    settingsRefused; // the keys of them it refused, as BoardSettings.h's bits
     uint32_t    recalibratedId;  // the last recalibration it ran; 0 for none
     uint16_t    recalibratedPpm;
     bool        recalibratedOk;
